[feat] 자기 자신의 비밀번호를 바꾸는 api 구현 (PUT / password)

+++ src/main/java/com/tmax/hyperauth/rest/PasswordProvider.java
@@ -3,12 +3,23 @@ package com.tmax.hyperauth.rest;
 import com.tmax.hyperauth.caller.StringUtil;
 import com.tmax.hyperauth.jpa.EmailVerification;
 import org.jboss.resteasy.spi.HttpResponse;
+import org.keycloak.OAuthErrorException;
+import org.keycloak.TokenVerifier;
 import org.keycloak.common.ClientConnection;
+import org.keycloak.common.VerificationException;
 import org.keycloak.connections.jpa.JpaConnectionProvider;
+import org.keycloak.crypto.SignatureProvider;
+import org.keycloak.crypto.SignatureVerifierContext;
+import org.keycloak.models.ClientModel;
 import org.keycloak.models.KeycloakSession;
+import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserCredentialModel;
 import org.keycloak.models.utils.FormMessage;
 import org.keycloak.models.utils.KeycloakModelUtils;
+import org.keycloak.protocol.oidc.TokenManager;
+import org.keycloak.representations.AccessToken;
+import org.keycloak.services.ErrorResponseException;
+import org.keycloak.services.Urls;
 import org.keycloak.services.resource.RealmResourceProvider;
 
 import javax.persistence.EntityManager;
@@ -44,6 +55,8 @@ public class PasswordProvider implements RealmResourceProvider {
     public Object getResource() {
         return this;
     }
+    private AccessToken token;
+    private ClientModel clientModel;
 
     Connection conn = null;
     long time = System.currentTimeMillis();
@@ -52,21 +65,42 @@ public class PasswordProvider implements RealmResourceProvider {
 
 	@PUT
     @Produces(MediaType.APPLICATION_JSON)
-    public Response get(@QueryParam("email") final String email, @QueryParam("code") String code, @QueryParam("password") String password, @QueryParam("confirmPassword") String confirmPassword) {
+    public Response get(@QueryParam("email") final String email, @QueryParam("code") String code ,@QueryParam("token") String tokenString,
+                        @QueryParam("password") String password, @QueryParam("confirmPassword") String confirmPassword) {
         System.out.println("***** PUT /password");
-        System.out.println("email : " + email);
-        boolean isVerified = false;
         try {
-            List< EmailVerification > emailCodeList = getEntityManager().createNamedQuery("findByEmail", EmailVerification.class)
-                    .setParameter("email", email).getResultList();
-            for ( EmailVerification emailCode : emailCodeList ){
-                if (code.equalsIgnoreCase(emailCode.getCode()) && emailCode.getIsVerified()) {
-                    isVerified = true;
-                    break;
+            if (StringUtil.isEmpty(email)){
+                status = Status.BAD_REQUEST;
+                out = "Email Empty";
+                return Util.setCors(status, out);
+            }
+            System.out.println("email : " + email);
+
+            RealmModel realm = session.realms().getRealmByName("tmax");
+            List< EmailVerification > emailCodeList = null;
+
+            boolean isVerified = false;
+
+            if ( StringUtil.isNotEmpty(code) && StringUtil.isEmpty(tokenString)){
+                emailCodeList = getEntityManager().createNamedQuery("findByEmail", EmailVerification.class)
+                        .setParameter("email", email).getResultList();
+                for ( EmailVerification emailCode : emailCodeList ){
+                    if (code.equalsIgnoreCase(emailCode.getCode()) && emailCode.getIsVerified()) {
+                        isVerified = true;
+                        break;
+                    }
                 }
+            } else if (StringUtil.isNotEmpty(tokenString) && StringUtil.isEmpty(code)){
+                verifyToken(tokenString, realm);
+                if (!token.getPreferredUsername().equalsIgnoreCase(email)) {
+                    out = "Cannot change other user's password";
+                    throw new ErrorResponseException(OAuthErrorException.INVALID_REQUEST, "Cannot change other user's password", Response.Status.BAD_REQUEST);
+                }
+                isVerified = true;
+            } else {
+
             }
             System.out.println("isVerified : " + isVerified);
-
             // Validation
             if (!isVerified) {
                 status = Status.UNAUTHORIZED;
@@ -108,13 +142,14 @@ public class PasswordProvider implements RealmResourceProvider {
             session.userCredentialManager().updateCredential(session.realms().getRealmByName("tmax"),
                     session.users().getUserByEmail(email, session.realms().getRealmByName("tmax")),
                     UserCredentialModel.password(password, false));
-
             System.out.println("Change Password Success");
 
-            for ( EmailVerification emailCode : emailCodeList ){
-                getEntityManager().remove(emailCode);
+            if ( StringUtil.isNotEmpty(code) && StringUtil.isEmpty(tokenString)){
+                for ( EmailVerification emailCode : emailCodeList ){
+                    getEntityManager().remove(emailCode);
+                }
+                System.out.println("delete DB data Success");
             }
-            System.out.println("delete DB data Success");
 
             status = Status.OK;
             out = "Reset Password Success";
@@ -143,5 +178,28 @@ public class PasswordProvider implements RealmResourceProvider {
 
     private EntityManager getEntityManager() {
         return session.getProvider(JpaConnectionProvider.class).getEntityManager();
+    }
+
+    private void verifyToken(String tokenString, RealmModel realm) throws VerificationException {
+        TokenVerifier<AccessToken> verifier = TokenVerifier.create(tokenString, AccessToken.class).withDefaultChecks()
+                .realmUrl(Urls.realmIssuer(session.getContext().getUri().getBaseUri(), realm.getName()));
+
+        SignatureVerifierContext verifierContext = session.getProvider(SignatureProvider.class,
+                verifier.getHeader().getAlgorithm().name()).verifier(verifier.getHeader().getKeyId());
+        verifier.verifierContext(verifierContext);
+        try {
+            token = verifier.verify().getToken();
+        } catch (Exception e) {
+            out = "token invalid";
+        }
+        clientModel = realm.getClientByClientId(token.getIssuedFor());
+        if (clientModel == null) {
+            out = "Client not found";
+            throw new ErrorResponseException(OAuthErrorException.INVALID_REQUEST, "Client not found", Status.NOT_FOUND);
+        }
+
+        TokenVerifier.createWithoutSignature(token)
+                .withChecks(TokenManager.NotBeforeCheck.forModel(clientModel))
+                .verify();
     }
 }
